+++ include/TsParser.h
@@ -129,21 +129,21 @@ public:
      * several TS-Packets for generating a complete PES-Packet. This function is used
      * internally by collectPes().
      */
-    void parsePesPacket(int16_t pid);
+    void parsePesPacket(int pid);
 
     /*!
      * Return raw bytes of table at pid
      */
-    ByteVector& getRawTable(int16_t pid);
+    ByteVector& getRawTable(int pid);
     TsStatistics mStatistics;
 
 private:
     // TODO maybe 1 parser per pid?
     std::map<int, ByteVector> mSectionBuffer;
-    std::map<uint16_t, int> mSectionLength;
-    std::map<uint16_t, int> mTableId;
-    std::map<uint16_t, int> mReadSectionLength;
-    std::map<uint16_t, PesPacket> mPesPacket;
+    std::map<int, int> mSectionLength;
+    std::map<int, int> mTableId;
+    std::map<int, int> mReadSectionLength;
+    std::map<int, PesPacket> mPesPacket;
 };
 
 } // namespace mpeg2ts
